% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_function.R
\name{kmeans_function}
\alias{kmeans_function}
\title{This function calculates the clusters from the given input
It is basically a wrapper around the kmeans function itself}
\usage{
kmeans_function(selectedData, clusters)
}
\arguments{
\item{selectedData}{The data from which the clusters and centers are going to be generated}

\item{clusters}{The number of clusters being used to plot \code{clusters}}
}
\value{
a cluster object with centers and clusters
}
\description{
This function calculates the clusters from the given input
It is basically a wrapper around the kmeans function itself
}
