% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_function.R
\name{knn_function}
\alias{knn_function}
\title{K-Nearest Neighbors}
\usage{
knn_function(selectedData, clusters)
}
\arguments{
\item{selectedData}{The data from which the clusters and centers are going to be generated}

\item{clusters}{The number of neighbors being used to plot \code{clusters}}
}
\value{
a cluster object with centers and clusters
}
\description{
This function calculates the nearest neighbors
}
