% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plot_function}
\alias{plot_function}
\title{This function plots the clusters from the output of the other function}
\usage{
plot_function(selectedDate, clusters)
}
\arguments{
\item{clusters}{The clusters calculated from the kmeans function}

\item{selectedData}{The data from which the clusters and centers are going to be generated}
}
\description{
Plotting the clusters and the centers
}
