% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R, R/plot_umap.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot Function PCA}
\usage{
plot_pca(umap_iris)

plot_pca(umap_iris)
}
\arguments{
\item{selectedData}{The data from which the clusters and centers are going to be generated}
}
\description{
This function plots the clusters from the output of the other function

This function plots the UMAP dimensionality reduction in two dimensions
}
\details{
Plotting the clusters and the centers
}
