% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_k_means.R
\name{shiny_k_means}
\alias{shiny_k_means}
\title{This function plots the clusters from the output of the other function}
\usage{
shiny_k_means()
}
\arguments{
\item{selectedData}{The data from which the clusters and centers are going to be generated}

\item{clusters}{The clusters calculated from the kmeans function}
}
\description{
Plotting the clusters and the centers
}
