% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_function.R
\name{umap_function}
\alias{umap_function}
\title{UMAP function}
\usage{
umap_function(selectedData)
}
\arguments{
\item{selectedData}{The data from which the clusters and centers are going to be generated}
}
\value{
a layput object is created
}
\description{
This function does the umap dimensional reduction
}
