% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plot_function}
\alias{plot_function}
\title{Plot Function Kmeans}
\usage{
plot_function(selectedDate, clusters)
}
\arguments{
\item{clusters}{The clusters calculated from the kmeans function}

\item{selectedData}{The data from which the clusters and centers are going to be generated}
}
\description{
This function plots the clusters from the output of the other function
}
\details{
Plotting the clusters and the centers
}
