% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_k_means.R
\name{shiny_k_means}
\alias{shiny_k_means}
\title{Shiny K means function}
\usage{
shiny_k_means()
}
\description{
This function will create the shiny app for the kmeans package.
The appdir object basically looks for the shiny folder in the kmeans
package and runs the ui and server functions.
}
\details{
The ui.R file will create the ui for the application.
The server.R will set up the server.
The app.R function sources both the ui and server functions.
}
